Add necessary .js files
