+++ src/turns.js
@@ -1,0 +1,9 @@
+function evaluateGuess(guess, answer) {
+    if (guess === answer) {
+        return 'Correct!'
+    } else {
+        return 'Incorrect!'
+    }
+}
+
+module.exports = { evaluateGuess }
