+++ test/deck-test.js
@@ -1,0 +1,4 @@
+const chai = require('chai');
+const expect = chai.expect;
+
+const {  } = require('../src/deck');
